requirements.txt
bleak~=0.14.2
numpy<2.0
opencv-python<5.0
Pillow>=8.0.0